pyproject.toml
[project]
name = "vv-clawbot"
version = "2.0.0"
description = "vv-clawbot: an agentic assistant with RAG, mem0 memory, and MCP tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.0",
    "slack-sdk>=3.27.0",
    "openai>=1.67.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "chromadb>=0.4.0",
    "mem0ai>=1.0.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.5",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
vv-clawbot = "clawbot_lib.main:run"

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawbot_lib"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "N816"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/clawbot_lib"]
omit = ["*/tests/*", "*/__pycache__/*"]
